refactor(rights) relocate RightsTemplateName class update module src.schemas.rights (from):     remove class RightsTemplateName update module src.utils.enums.rights (to):     add class RightsTemplateName
